Consolidate templating into RenderingService with dependency injection
Remove lazy loading DI violations by:

1. VisibilityService: Remove dead singleton infrastructure (init_visibility_service,
   get_visibility_service, etc.) - use direct instantiation instead

2. Create RenderingService that consolidates:
   - mudd/templating/ (TemplateRenderer, TemplateRenderError)
   - mudd/formatting/entities.py (build_contents_string, format_entity_with_contents,
     format_room_entities, render_entity_on_look)

3. Inject RenderingService into Look, Interact, and Sync cogs
   - Sync cog calls clear_cache() after entity sync

4. Delete consolidated modules:
   - mudd/templating/
   - mudd/formatting/
