+++ main.py
@@ -15,7 +15,8 @@ from mudd.cogs.sync import Sync
 from mudd.services.database import close_pool, get_pool, init_database
 from mudd.services.entity import EntityService
 from mudd.services.focus_context import FocusContextService
-from mudd.services.visibility import init_visibility_service
+from mudd.services.rendering import RenderingService
+from mudd.services.visibility import VisibilityService
 
 load_dotenv()
 
@@ -69,16 +70,28 @@ async def setup_hook():
     # Create services with explicit dependencies
     entity_service = EntityService(pool)
     focus_service = FocusContextService(pool)
-    visibility_service = init_visibility_service(pool)
+    visibility_service = VisibilityService(pool)
+    rendering_service = RenderingService()
 
     # Create cogs with explicit dependencies
     await bot.add_cog(
-        Interact(bot, entity_service, focus_service, visibility_service, pool)
+        Interact(
+            bot,
+            entity_service,
+            focus_service,
+            visibility_service,
+            pool,
+            rendering_service,
+        )
     )
-    await bot.add_cog(Look(bot, entity_service, focus_service, visibility_service))
+    await bot.add_cog(
+        Look(bot, entity_service, focus_service, visibility_service, rendering_service)
+    )
     await bot.add_cog(Ping(bot))
     await bot.add_cog(Movement(bot, visibility_service, focus_service))
-    await bot.add_cog(Sync(bot, entity_service, visibility_service, pool))
+    await bot.add_cog(
+        Sync(bot, entity_service, visibility_service, pool, rendering_service)
+    )
 
 
 @bot.event
