--- mudd/formatting/__init__.py
@@ -1,1 +1,0 @@
-"""Formatting utilities for MUDD output."""
