+++ tests/unit/test_template_renderer.py
@@ -3,8 +3,11 @@
 import pytest
 
 from mudd.services.entity import ResolvedEntity
-from mudd.templating import TemplateRenderError, clear_cache, render
-from mudd.templating.renderer import TemplateRenderer
+from mudd.services.rendering import (
+    RenderingService,
+    TemplateRenderer,
+    TemplateRenderError,
+)
 
 
 def make_entity(
@@ -32,72 +35,80 @@ def make_entity(
     )
 
 
-class TestRender:
-    """Tests for public render() function."""
+@pytest.fixture
+def rendering_service() -> RenderingService:
+    """Create a RenderingService for tests."""
+    return RenderingService()
 
-    def test_simple_template_renders(self):
+
+class TestRender:
+    """Tests for RenderingService.render() method."""
+
+    def test_simple_template_renders(self, rendering_service):
         """Simple template text renders correctly."""
         entity = make_entity()
-        result = render("Hello world", entity)
+        result = rendering_service.render("Hello world", entity)
         assert result == "Hello world"
 
-    def test_template_with_name_variable(self):
+    def test_template_with_name_variable(self, rendering_service):
         """Template can use {{ name }} variable."""
         entity = make_entity()
-        result = render("You look at the {{ name }}.", entity)
+        result = rendering_service.render("You look at the {{ name }}.", entity)
         assert result == "You look at the *Test Entity*."
 
-    def test_template_with_entity_properties(self):
+    def test_template_with_entity_properties(self, rendering_service):
         """Template can access entity properties via e."""
         entity = make_entity(description_long="A sturdy table.")
-        result = render("{{ e.description_long }}", entity)
+        result = rendering_service.render("{{ e.description_long }}", entity)
         assert result == "A sturdy table."
 
-    def test_template_with_conditional(self):
+    def test_template_with_conditional(self, rendering_service):
         """Template can use Jinja conditionals."""
         entity = make_entity(description_long="Detailed text.")
         template = (
             "{% if e.description_long %}{{ e.description_long }}"
             "{% else %}Nothing special.{% endif %}"
         )
-        result = render(template, entity)
+        result = rendering_service.render(template, entity)
         assert result == "Detailed text."
 
-    def test_template_with_or_fallback(self):
+    def test_template_with_or_fallback(self, rendering_service):
         """Template can use or for fallback values."""
         entity = make_entity(description_long=None, description_short="Short desc.")
         template = "{{ e.description_long or e.description_short or 'Nothing.' }}"
-        result = render(template, entity)
+        result = rendering_service.render(template, entity)
         assert result == "Short desc."
 
-    def test_none_template_returns_empty_string(self):
+    def test_none_template_returns_empty_string(self, rendering_service):
         """None template returns empty string."""
         entity = make_entity()
-        result = render(None, entity)
+        result = rendering_service.render(None, entity)
         assert result == ""
 
-    def test_syntax_error_raises_exception(self):
+    def test_syntax_error_raises_exception(self, rendering_service):
         """Template syntax error raises TemplateRenderError."""
         entity = make_entity()
         with pytest.raises(TemplateRenderError):
-            render("{% if %}broken{% endif %}", entity)
+            rendering_service.render("{% if %}broken{% endif %}", entity)
 
-    def test_undefined_variable_raises_exception(self):
+    def test_undefined_variable_raises_exception(self, rendering_service):
         """Undefined variable raises TemplateRenderError."""
         entity = make_entity()
         with pytest.raises(TemplateRenderError):
-            render("{{ undefined_var }}", entity)
+            rendering_service.render("{{ undefined_var }}", entity)
 
-    def test_contents_variable_available(self):
+    def test_contents_variable_available(self, rendering_service):
         """Template can use {{ contents }} variable."""
         entity = make_entity()
-        result = render("Items:{{ contents }}", entity, contents="\n- a *Vase*")
+        result = rendering_service.render(
+            "Items:{{ contents }}", entity, contents="\n- a *Vase*"
+        )
         assert result == "Items:\n- a *Vase*"
 
-    def test_contents_defaults_to_empty_string(self):
+    def test_contents_defaults_to_empty_string(self, rendering_service):
         """{{ contents }} is empty string when not provided."""
         entity = make_entity()
-        result = render("[{{ contents }}]", entity)
+        result = rendering_service.render("[{{ contents }}]", entity)
         assert result == "[]"
 
 
@@ -110,11 +121,11 @@ class TestTemplateRenderer:
         entity = make_entity()
         template_source = "Cached template {{ name }}"
 
-        renderer.render(template_source, entity, "*Test*")
+        renderer.render_template(template_source, entity, "*Test*")
         assert template_source in renderer._cache
 
         # Second render uses cache
-        renderer.render(template_source, entity, "*Test2*")
+        renderer.render_template(template_source, entity, "*Test2*")
         assert len(renderer._cache) == 1
 
     def test_clear_cache(self):
@@ -122,24 +133,24 @@ class TestTemplateRenderer:
         renderer = TemplateRenderer()
         entity = make_entity()
 
-        renderer.render("Template 1", entity, "*Test*")
-        renderer.render("Template 2", entity, "*Test*")
+        renderer.render_template("Template 1", entity, "*Test*")
+        renderer.render_template("Template 2", entity, "*Test*")
         assert len(renderer._cache) == 2
 
         renderer.clear_cache()
         assert len(renderer._cache) == 0
 
 
-class TestClearCache:
-    """Tests for public clear_cache() function."""
+class TestRenderingServiceClearCache:
+    """Tests for RenderingService.clear_cache() method."""
 
-    def test_clear_cache_function(self):
-        """clear_cache() clears the singleton renderer's cache."""
+    def test_clear_cache_method(self, rendering_service):
+        """clear_cache() clears the renderer's template cache."""
         entity = make_entity()
 
         # Render some templates to populate cache
-        render("Template A", entity)
-        render("Template B", entity)
+        rendering_service.render("Template A", entity)
+        rendering_service.render("Template B", entity)
 
         # Clear and verify no exception
-        clear_cache()
+        rendering_service.clear_cache()
